requirements.txt
openai
chromadb
python-dotenv
gradio
tqdm
tenacity
litellm